requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0

urllib3
pyyaml